feat: Refactor user authentication and authorization
This commit refactors the user authentication and authorization functionality. It introduces a new utility function, `generateToken`, for JWT token generation. The `createUser` route now generates a token upon successful user creation, and the `loginUser` route generates a token upon successful login. The `logoutCurrentUser` route is added to handle user logout. The `authMiddleware` file is removed as it is no longer needed. Additionally, the `categoryController`, `categoryModel`, and related routes are removed as they are no longer used.