feat: Add deleteUserById route and controller function
The code changes add a new route `deleteUserById` in the `userRoutes.js` file to delete a user by their ID. The corresponding controller function `deleteUserById` is also added in the `userController.js` file. This commit enhances the functionality of the application by allowing administrators to delete users from the system.